+++ src/main/java/io/github/ladysnake/impersonate/impl/mixin/ServerPlayNetworkHandlerMixin.java
@@ -21,6 +21,7 @@ import io.github.ladysnake.impersonate.Impersonator;
 import io.github.ladysnake.impersonate.impl.PacketMeddling;
 import io.github.ladysnake.impersonate.impl.RecipientAwareText;
 import net.minecraft.network.Packet;
+import net.minecraft.network.packet.s2c.play.ChatMessageS2CPacket;
 import net.minecraft.network.packet.s2c.play.GameMessageS2CPacket;
 import net.minecraft.network.packet.s2c.play.PlayerListS2CPacket;
 import net.minecraft.server.network.ServerPlayNetworkHandler;
@@ -39,7 +40,11 @@ public abstract class ServerPlayNetworkHandlerMixin {
 
     @ModifyArg(method = "sendPacket(Lnet/minecraft/network/Packet;Lnet/minecraft/network/PacketCallbacks;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/network/ClientConnection;send(Lnet/minecraft/network/Packet;Lnet/minecraft/network/PacketCallbacks;)V"))
     private Packet<?> resolveFakeTextsInPackets(Packet<?> packet) {
-        if (packet instanceof GameMessageS2CPacket gamePacket) {
+        if (packet instanceof ChatMessageS2CPacket chatPacket) {
+            if (this.existsImpersonator()) {
+                return PacketMeddling.resolveChatMessage(chatPacket, player);
+            }
+        } else if (packet instanceof GameMessageS2CPacket gamePacket) {
             if (this.existsImpersonator()) {
                 Text resolvedText = ((RecipientAwareText) gamePacket.content()).impersonateResolveAll(player);
                 return new GameMessageS2CPacket(resolvedText, gamePacket.overlay());